+++ client_connection/manager.go
@@ -1,55 +1,27 @@
 package client_connection
 
-import (
-	"errors"
-	"github.com/mysterium/node/service_discovery/dto"
-)
+import "github.com/mysterium/node/service_discovery/dto"
 
-type State string
-
-const (
-	NOT_CONNECTED = State("NOT_CONNECTED")
-	NEGOTIATING   = State("NEGOTIATING")
-	CONNECTED     = State("CONNECTED")
-)
-
-var (
-	ALREADY_CONNECTED = errors.New("already connected")
-)
-
-type ConnectionStatus struct {
-	State     State
-	SessionId string
+type vpnManager struct {
 }
 
-type Manager interface {
-	Connect(identity dto.Identity, NodeKey string) error
-	Status() ConnectionStatus
-	Disconnect() error
-	Wait() error
+func NewVpnManager() *vpnManager {
+
+	return &vpnManager{}
 }
 
-type fakeManager struct {
-	errorChannel chan error
-}
-
-func NewManager() *fakeManager {
-	return &fakeManager{make(chan error)}
-}
-
-func (nm *fakeManager) Connect(identity dto.Identity, NodeKey string) error {
+func (vpn *vpnManager) Connect(identity dto.Identity, NodeKey string) error {
 	return nil
 }
 
-func (nm *fakeManager) Status() ConnectionStatus {
-	return ConnectionStatus{NOT_CONNECTED, ""}
+func (vpn *vpnManager) Status() ConnectionStatus {
+	return ConnectionStatus{}
 }
 
-func (nm *fakeManager) Disconnect() error {
-	nm.errorChannel <- errors.New("disconnected")
+func (vpn *vpnManager) Disconnect() error {
 	return nil
 }
 
-func (nm *fakeManager) Wait() error {
-	return <-nm.errorChannel
+func (vpn *vpnManager) Wait() error {
+	return nil
 }
